chore: SP-1247 Add all concluded and detected licenses to SPDXLite export
