+++ src/main/task/export/Export.ts
@@ -3,12 +3,10 @@
 import { NewExportDTO, SourceType } from 'api/dto';
 import { Format } from '../../modules/export/Format';
 import { Spdxv20 } from '../../modules/export/format/Spdxv20';
-import { SpdxLite } from '../../modules/export/format/SpdxLite';
 import { Csv } from '../../modules/export/format/Csv';
 import { Raw } from '../../modules/export/format/Raw';
 import { Wfp } from '../../modules/export/format/Wfp';
 import { HtmlSummary } from '../../modules/export/format/HtmlSummary';
-import { SpdxLiteJson } from '../../modules/export/format/SpdxLiteJson';
 import { ITask } from '../Task';
 import { IExportResult } from '../../modules/export/IExportResult';
 import { ExportFormat, ExportSource, InventoryType } from '../../../api/types';
@@ -17,6 +15,8 @@ import { workspace } from '../../workspace/Workspace';
 import { ExportModel } from '../../modules/export/Model/ExportModel';
 import { CycloneDXDetected } from '../../modules/export/format/CycloneDX/CycloneDXDetected';
 import { CycloneDXIdentified } from '../../modules/export/format/CycloneDX/CycloneDxIdentified';
+import { SpdxLiteDetected } from '../../modules/export/format/SPDXLite/SpdxLiteDetected';
+import { SpdxLiteIdentified } from '../../modules/export/format/SPDXLite/SpdxLiteIdentified';
 
 export class Export implements ITask<string, IExportResult> {
   private format: Format;
@@ -36,9 +36,6 @@ export class Export implements ITask<string, IExportResult> {
       case ExportFormat.SPDX20:
         this.format = new Spdxv20();
         break;
-      case ExportFormat.SPDXLITE:
-        this.format = new SpdxLite(exportDTO.source);
-        break;
       case ExportFormat.CSV:
         if (exportDTO.inventoryType === InventoryType.SBOM) {
           this.format = new Csv(exportDTO.source);
@@ -54,10 +51,10 @@ export class Export implements ITask<string, IExportResult> {
         this.format = new Wfp();
         break;
       case ExportFormat.SPDXLITEJSON:
-        this.format = new SpdxLiteJson(exportDTO.source);
+        this.format = exportDTO.source === ExportSource.DETECTED ? new SpdxLiteDetected(exportDTO.source) : new SpdxLiteIdentified(exportDTO.source);
         break;
       case ExportFormat.CYCLONEDX:
-        this.format = exportDTO.source === ExportSource.DETECTED ? new CycloneDXDetected(exportDTO.source, workspace.getOpenedProjects()[0], new ExportModel()) : new CycloneDXIdentified(exportDTO.source, workspace.getOpenedProjects()[0], new ExportModel()) ;
+        this.format = exportDTO.source === ExportSource.DETECTED ? new CycloneDXDetected(exportDTO.source, workspace.getOpenedProjects()[0], new ExportModel()) : new CycloneDXIdentified(exportDTO.source, workspace.getOpenedProjects()[0], new ExportModel());
         break;
       case ExportFormat.HTMLSUMMARY:
         this.format = new HtmlSummary(exportDTO.source);
